+++ frontend/src/visualization/charts_option_xxs.tsx
